feat(workout-cards): put workout-cards in component
