+++ src/pages/index.jsx
@@ -2,6 +2,7 @@ import React from "react";
 import Image from "next/image";
 import Button from "../atoms/button";
 import styled from "@emotion/styled";
+import CardWorkout from "../molecules/card-workout";
 
 // image link:
 // https://images.unsplash.com/photo-1514306191717-452ec28c7814?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1169&q=80
@@ -17,21 +18,6 @@ const B = styled(A)`
 	background: pink;
 `;
 
-const exercises = [
-	{
-		id: "Clown",
-		name: "Clown",
-		description: "This is a clown exercise",
-		duration: "5min", // 5 minutes -> 5*60*1000
-	},
-	{
-		id: "Fixpoint",
-		name: "Fixpoint",
-		description: "This is a Fixpoint exercise",
-		duration: "10min", // 10 minutes -> 10*60*1000
-	},
-];
-
 const Page = () => {
 	return (
 		<B>
@@ -44,16 +30,8 @@ const Page = () => {
 				layout="fixed"
 			/>
 			<Button>Mime Workout</Button>
-			<ul>
-				{exercises.map(exercise => {
-					return (
-						<li key={exercise.id}>
-							<h2>{exercise.name}</h2>
-							<p>{exercise.description}</p>
-						</li>
-					);
-				})}
-			</ul>
+
+			<CardWorkout />
 		</B>
 	);
 };
